requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
pertdist==0.1.0
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.0
